% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{format_deparsed}
\alias{format_deparsed}
\title{format deparsed code, removing brackets and starts of lines if necessary}
\usage{
format_deparsed(deparsed)
}
\arguments{
\item{deparsed}{A character vector from a use of deparse}
}
\description{
format deparsed code, removing brackets and starts of lines if necessary
}

