% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionally.R
\name{optionally}
\alias{optionally}
\title{Consult an option, then default}
\description{
Arguments that appear like so in the usage:\preformatted{f(..., arg = optionally(DEFAULT), ...)
}

get their value according to this logic:\preformatted{user-specified value or, if not given,
  getOption("pkg.arg") or if does not exist,
    DEFAULT
}

The user can provide a value in the call. Otherwise, an option is consulted.
If that does not exist, then the documented default is used. This is not an
exported function and should not be called directly.
}
