% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex-package.R
\docType{package}
\name{reprex-package}
\alias{reprex-package}
\alias{_PACKAGE}
\title{reprex: Prepare Reproducible Example Code via the Clipboard}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Convenience wrapper that uses the 'rmarkdown'
    package to render small snippets of code to target formats that
    include both code and output.  The goal is to encourage the sharing of
    small, reproducible, and runnable examples on code-oriented websites,
    such as <https://stackoverflow.com> and <https://github.com>, or in
    email. The user's clipboard is the default source of input code and
    the default target for rendered output. 'reprex' also extracts clean,
    runnable R code from various common formats, such as copy/paste from
    an R session.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reprex.tidyverse.org}
  \item \url{https://github.com/tidyverse/reprex#readme}
  \item Report bugs at \url{https://github.com/tidyverse/reprex/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@rstudio.com} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Authors:
\itemize{
  \item Jim Hester \email{james.hester@rstudio.com} (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})
  \item David Robinson \email{admiral.david@gmail.com}
  \item Hadley Wickham \email{hadley@rstudio.com} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
