% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex-undo.R
\name{reprex_clean}
\alias{reprex_clean}
\title{Delete commented output from a reprex}
\usage{
reprex_clean(x = NULL, comment = "^#>")
}
\arguments{
\item{x}{character, holding the lines of a rendered or displayed reprex. If
not provided, the clipboard is consulted for input.}

\item{comment}{regular expression that matches commented output lines}
}
\value{
character vector holding just the clean R code, invisibly
}
\description{
Cleans commented output out of a displayed reprex, such as code copied from a
GitHub issue. The clean reprex code is printed, returned invisibly, and
written to the clipboard, if possible.
}
\examples{
x <- c(
  "## a regular comment, which is retained",
  "(x <- 1:4)",
  "#> [1] 1 2 3 4",
  "median(x)",
  "#> [1] 2.5"
  )
reprex_clean(x)
}
\seealso{
\code{\link{reprex_invert}()} if the input is actual Markdown, i.e.
  the direct output of \code{\link{reprex}()}.
}
