% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex_document.R
\name{reprex_document}
\alias{reprex_document}
\title{reprex document format}
\usage{
reprex_document(pandoc_args = NULL)
}
\arguments{
\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\value{
something
}
\description{
reprex document format
}
