% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex-info.R
\name{reprex_info}
\alias{reprex_info}
\title{Report when and how reprex was made}
\usage{
reprex_info()
}
\value{
String documenting when reprex was rendered and how
}
\description{
Report when and how reprex was made
}
\examples{
reprex_info()
}
