% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex-locale.R
\name{reprex_locale}
\alias{reprex_locale}
\title{Render a reprex in a different locale}
\usage{
reprex_locale(..., locale)
}
\arguments{
\item{...}{Parameters passed onto \code{\link[=reprex]{reprex()}}.}

\item{locale}{String. The locale to be used when rendering the \code{\link[=reprex]{reprex()}}.}
}
\value{
Character vector of rendered reprex, invisibly.
}
\description{
Render a \code{\link[=reprex]{reprex()}} using a different locale.
}
\examples{
\dontrun{

# Rendering in Italian:
reprex_locale({
  dplyr::select(foo)
}, locale = "it_IT")

# Rendering in Spanish:
reprex_locale({
  df <- data.frame(
    stringsAsFactors = FALSE,
    date = as.Date(c("2019-01-01", "2019-02-01")),
    value = c(1, 2)
  )
  plot(df$date, df$value)
}, locale = "es_ES")
}

}
