% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venues.R
\name{reprex_venue}
\alias{reprex_venue}
\alias{reprex_html}
\alias{reprex_r}
\alias{reprex_rtf}
\alias{reprex_slack}
\title{Venue-specific shortcuts}
\usage{
reprex_html(...)

reprex_r(...)

reprex_rtf(...)

reprex_slack(...)
}
\arguments{
\item{...}{Passed along to \code{\link[=reprex]{reprex()}}.}
}
\description{
These are thin wrappers around \code{reprex()} that incorporate the target \code{venue}
as a suffix in the function name, for easier access via auto-completion.
}
