% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionally.R
\name{opt}
\alias{opt}
\title{Consult an option, then default}
\description{
Arguments that appear like so in the usage:\preformatted{f(..., arg = opt(DEFAULT), ...)
}

get their value according to this logic:\preformatted{user-specified value or, if not given,
  getOption("reprex.arg") or if does not exist,
    DEFAULT
}

It's shorthand for:\preformatted{f(..., arg = getOption("reprex.arg", DEFAULT), ...)
}

This is not an exported function and should not be called directly.
}
