% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex-addin.R
\name{reprex_addin}
\alias{reprex_addin}
\title{Render a reprex}
\usage{
reprex_addin()
}
\description{
An \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{http://rstudio.github.io/rstudioaddins/}{addin} to call
\code{\link{reprex}()}. Meant to be activated from the RStudio Addins menu.
Prepare in one of these ways:
\enumerate{
\item Copy reprex source to clipboard.
\item Select reprex source.
\item Activate the file containing reprex source.
}
Call \code{\link{reprex}()} directly for more control via additional
arguments.
}
