% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex-info.R
\name{reprex_info}
\alias{reprex_info}
\title{Report when and how reprex was made}
\usage{
reprex_info()
}
\value{
Character
}
\description{
\code{reprex_info()} returns a string documenting when the reprex was rendered and
how. It is included as the first line of reprex source when \code{advertise = TRUE}.
}
\examples{
reprex_info()
}
