% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex-undo.R
\name{reprex_invert}
\alias{reprex_invert}
\title{Un-render a reprex}
\usage{
reprex_invert(x = NULL, comment = "^#>")
}
\arguments{
\item{x}{character, holding the lines of a rendered or displayed reprex. If
not provided, the clipboard is consulted for input.}

\item{comment}{regular expression that matches commented output lines}
}
\value{
character vector holding just the clean R code, invisibly
}
\description{
A crude function to recover the input of \code{\link{reprex}()} from its
output, i.e. Markdown in a character vector.
}
\examples{
x <- reprex({
  #' Some text
  #+ chunk-label-and-options-cannot-be-recovered, message = TRUE
  (x <- 1:4)
  #' More text
  y <- 2:5
  x + y
}, show = FALSE)
x
reprex_invert(x)
}
\seealso{
\code{\link{reprex_clean}()} if the input is not Markdown, i.e. if
  it's just a rendered code chunk copied from a GitHub issue.
}
